test(cli): merge dynamic-import-graph specs into create-dependency-graph
